+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Fixed
+
+
+
+
 ## [0.12.0] - 2023-01-21
 
 ### Added
@@ -896,6 +907,7 @@ in most cases.
 This is the first release of Plom, PaperLess Open Marking.
 
 
+[Unreleased]: https://gitlab.com/plom/plom/compare/v0.12.0...main
 [0.12.0]: https://gitlab.com/plom/plom/-/compare/v0.11.1...v0.12.0
 [0.11.1]: https://gitlab.com/plom/plom/-/compare/v0.11.0...v0.11.1
 [0.11.0]: https://gitlab.com/plom/plom/-/compare/v0.10.1...v0.11.0